requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.8
django-filter==2.0.0
django-suit==0.2.26
django-webpack-loader==0.6.0
djangorestframework==3.8.2
flake8==3.5.0
gunicorn==19.9.0
idna==2.7
Markdown==2.6.11
mccabe==0.6.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
PyYAML==4.2b1
requests==2.20.0
urllib3==1.24.2
whitenoise==4.0
